made uri type one function with enum use
