+++ lib/types/uri-type.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const URI_TYPE = Object.freeze({
+  TRACK: 0,
+  LINE_IN: 1,
+  RADIO: 2
+});
+
+module.exports = URI_TYPE;
